Remove TPC only ERepSimTPCdef.hxx files.
The information in this file has been unified with the rest of erep-sim.
The TPC identifiers are now defined in ERepSimDefs, and made available to
the TPC simulation through the constructor.  The volume and selection names
are also made available through the constructors so that the TPC classes are
independent of the values.