Change the configuration methodology.
The earlier code put configuration values (like the detector volume names
into include files, but it was scatter into different places.  This collects
(most) of the information so that values are provided to the class
constructors.  This solves some nameing collisions in the properties, and
is going to make it easier to implement a configuration file.