Remove pointer references to hit segments
This no longer references the hit segment being used according to the
pointer to the hit segment.  The hit segment is copied by value into
the carrier, and it is identified according to it's unique
identification number.  This allows segments to be used after the
event is removed from memory.